+++ src/mako/error/handlers/web/DevelopmentHandler.php
@@ -73,19 +73,29 @@ class DevelopmentHandler extends Handler implements HandlerInterface
 	protected $app;
 
 	/**
+	 * Cookies and headers to keep.
+	 *
+	 * @var array
+	 */
+	protected $keep;
+
+	/**
 	 * Constructor.
 	 *
 	 * @param \mako\http\Request            $request  Request
 	 * @param \mako\http\Response           $response Response
 	 * @param \mako\application\Application $app      Application
+	 * @param \array                        $keep     Cookies and headers to keep
 	 */
-	public function __construct(Request $request, Response $response, Application $app)
+	public function __construct(Request $request, Response $response, Application $app, array $keep = [])
 	{
 		$this->request = $request;
 
 		$this->response = $response;
 
 		$this->app = $app;
+
+		$this->keep = $keep;
 	}
 
 	/**
@@ -467,7 +477,7 @@ class DevelopmentHandler extends Handler implements HandlerInterface
 		['type' => $type, 'body' => $body] = $this->buildResponse($exception);
 
 		$this->sendResponse($this->response
-		->clear()
+		->clearExcept($this->keep)
 		->disableCaching()
 		->disableCompression()
 		->setType($type)